requirements.txt
pip
bumpversion==0.5.3
flake8
tox==2.3.1
Sphinx
coverage
python-dotenv>=0.5.1
pytest>=2.9.2
sqlalchemy
